requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.62
botocore==1.31.62
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-redis==5.4.0
django-storages==1.14.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
rpds-py==0.12.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.6
